billing-service/pyproject.toml
[project]
name = "billing-service"
version = "0.1.0"
description = ""
authors = [
    {name = "BENHIMA Mohamed-amine",email = "benhima.mohamed.amine@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.127.0,<0.128.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-consul (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"